requirements.txt
django 3.1.2 (rootz)
django 3.1.3 (raspberry pi)
RPi.gpiozero
pigpio
RPIO